\name{NAdens-package}
\alias{NAdens-package}
\docType{package}
\title{
Functions for plotting data values and missing data frequencies together.
}
\description{
Plotting data frequencies and densities is handy for exploratory analysis, or just checking to see whether a data transformation worked as expected. This package provides functions to check the relative frequencies of missing data and the nonmissing data frequencies or density together.
}
\details{
\tabular{ll}{
Package: \tab NAdens\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2015-03-04\cr
License: \tab GPL>=2\cr
}
}
\author{
Nick Salkowski

Maintainer: Nick Salkowski <nicksalkowskiphd@gmail.com>
}
\keyword{ package }
