% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NAdens.character.R
\name{NAdens.character}
\alias{NAdens.character}
\title{Construct a \code{\link{NAdens}} class object from a character vector}
\usage{
\method{NAdens}{character}(x)
}
\arguments{
\item{x}{a character vector.}
}
\description{
Construct a \code{\link{NAdens}} class object from a character vector
}
\examples{
X <- c(sample(LETTERS[1:5], size = 40, replace = TRUE), rep(NA, 10))
Xdens <- NAdens(X)
plot(Xdens)
}

