% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NAdens.factor.R
\name{NAdens.factor}
\alias{NAdens.factor}
\title{Construct a \code{\link{NAdens}} class object from a factor vector}
\usage{
\method{NAdens}{factor}(x)
}
\arguments{
\item{x}{a factor vector.}
}
\description{
Construct a \code{\link{NAdens}} class object from a factor vector
}
\examples{
X <- factor(c(sample(LETTERS[1:5], size = 40, replace = TRUE), rep(NA, 10)))
Xdens <- NAdens(X)
plot(Xdens)
}

