% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NAdens.logical.R
\name{NAdens.logical}
\alias{NAdens.logical}
\title{Construct a \code{\link{NAdens}} class object from a logical vector}
\usage{
\method{NAdens}{logical}(x)
}
\arguments{
\item{x}{a logical vector.}
}
\description{
Construct a \code{\link{NAdens}} class object from a logical vector
}
\examples{
X <- c(sample(c(TRUE, FALSE), size = 40, replace = TRUE), rep(NA, 10))
Xdens <- NAdens(X)
plot(Xdens)
}

