% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NAdens.numeric.R
\name{NAdens.numeric}
\alias{NAdens.numeric}
\title{Construct a \code{\link{NAdens}} class object from a numeric vector}
\usage{
\method{NAdens}{numeric}(x, ...)
}
\arguments{
\item{x}{a character vector.}

\item{...}{Additional arguments passed to \code{\link[stats]{density}}.}
}
\description{
Construct a \code{\link{NAdens}} class object from a numeric vector
}
\examples{
X <- c(rnorm(40), rep(NA, 10))
Xdens <- NAdens(X)
plot(Xdens)
}

