% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.NAdens.R
\name{plot.NAdens}
\alias{plot.NAdens}
\title{Plot a \code{\link{NAdens}} class object}
\usage{
\method{plot}{NAdens}(dens, NA.label = "NA", col = NULL,
  border = par("fg"), main = NULL, xlab = NULL, ylab = NULL,
  bty = "n", ...)
}
\arguments{
\item{dens}{An \code{NAdens} class object.}

\item{NA.label}{Character label for \code{NA} values on the plot.}

\item{col}{Fill color.}

\item{border}{Border color.}

\item{main}{Plot title.}

\item{xlab}{X axis label.}

\item{ylab}{Y axis label.}

\item{bty}{Type of box to draw around the plot.}

\item{...}{Additional arguments to pass to \code{\link[graphics]{plot}}}
}
\description{
Plot a \code{\link{NAdens}} class object.
}
\examples{
X <- factor(c(sample(LETTERS[1:5], size = 40, replace = TRUE), rep(NA, 10)))
Xdens <- NAdens(X)
plot(Xdens)

XX <- c(rnorm(400), rep(NA, 100))
XXdens <- NAdens(XX)
plot(XXdens)
}

